Started persistent question/answer command, improved call creation/removal
Added the ability to store a question+answer+image object. Its stored in a json file so that it is persistent. The idea is to have the ability to create questions to answer like a quiz.

>create command renamed to >call.

Now checks if there is a category called 'Voice Channels' (the default discord category for voice channels, and sets that as the parent for created calls.

Call removal now makes sure its a voice channel before removing. Before it would remove categories, text channels, etc.