Rainbow won't work I believe
Works fine for a few minutes but eventually starts slowing down to a crawl. Probably related to the discord rate limits so it likely works if it's set to a very high interval but at that point I'm not sure it's worth having at all. Tried on 100 - 500ms interval with eventual problems. I imagine anywhere above 30 seconds would be fine, just not cool.